* commit 'ef4c8a444d6558d18d3e453f7a4d080953c7c3a2 [formerly 46c440344e864d1368a2a15b2036b72f1f521b20]':
  ODY-1443: valueOf requires String! added file properties again


Former-commit-id: 4f5403df817471e1a8160f29bdfec022d9e19763

+++ common/mep/src/de/uniol/inf/is/odysseus/mep/functions/string/SubStringFunction2.java
@@ -37,11 +37,12 @@ public class SubStringFunction2 extends AbstractFunction<String> {
 	@Override
 	public String getValue() {
 		String a = getInputValue(0);
-		Integer b = Integer.valueOf(getInputValue(1));
+		Long b = getInputValue(1);
+
 		if ((a == null) || (b == null)) {
 			return null;
 		}
-		return a.substring(b);
+		return a.substring(b.intValue());
 	}
 
 }
